requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
langchain-openai>=0.0.5
mcp-use>=1.1.5 